Test.StateMachine.BoxDrawer

Description

This module contains functions for visualing a history of a parallel execution.

Documentation

data EventType

Event invocation or response.

Constructors

Open

Close

data Fork a

Constructors

Fork a a a

exec ∷ [(EventType, Pid)] → Fork [String] → Doc

Given a history, and output from processes generate Doc with boxes