Test.StateMachine.DotDrawing

Documentation

data GraphOptions

Constructors

GraphOptions

Fields

filePath ∷ FilePath

graphvizOutput ∷ GraphvizOutput

data GraphvizOutput

The possible Graphviz output formats (that is, those that actually produce a file).

Constructors

Bmp
Windows Bitmap Format.

Canon
Pretty-printed Dot output with no layout performed.

DotOutput
Reproduces the input along with layout information.

XDot (Maybe Version)
As with DotOutput, but provides even more information on how the graph is drawn. The optional Version is the same as specifying the XDotVersion attribute.

Eps
Encapsulated PostScript.

Fig
FIG graphics language.

Gd
Internal GD library format.

Gd2
Compressed version of Gd.

Gif
Graphics Interchange Format.

Ico
Icon image file format.

Imap
Server-side imagemap.

Cmapx
Client-side imagemap.

ImapNP
As for Imap, except only rectangles are used as active areas.

CmapxNP
As for Cmapx, except only rectangles are used as active areas.

Jpeg
The JPEG image format.

Pdf
Portable Document Format.

Plain
Simple text format.

PlainExt
As for Plain, but provides port names on head and tail nodes when applicable.

Png
Portable Network Graphics format.

Ps
PostScript.

Ps2
PostScript for PDF.

Svg
Scalable Vector Graphics format.

SvgZ
Compressed SVG format.

Tiff
Tagged Image File Format.

Vml
Vector Markup Language; Svg is usually preferred.

VmlZ
Compressed VML format; SvgZ is usually preferred.

Vrml
Virtual Reality Modeling Language format; requires nodes to have a third dimension set via the Pos attribute (and with a Dim value of at least 3).

WBmp
Wireless BitMap format; monochrome format usually used for mobile computing devices.

WebP
Google's WebP format; requires Graphviz >= 2.29.0.

data Rose a

Constructors

Rose a (Map Pid a)

printDotGraph ∷ GraphOptions → Rose [String] → IO ()