Test.StateMachine.Labelling

Description

This module exports helpers that are useful for labelling properties.

Documentation

data Predicate a b

Constructors

Predicate

Fields

predApply ∷ a → Either b (Predicate a b)

Given an a, either successfully classify as b or continue looking

predFinish ∷ Maybe b

End of the string

The predicate is given a final chance to return a value.

predicate ∷ (a → Either b (Predicate a b)) → Predicate a b

Construct simply predicate that returns Nothing on termination

maximum ∷ ∀ a b. Ord b ⇒ (a → Maybe b) → Predicate a b

Maximum value found, if any

classify ∷ ∀ a b. [Predicate a b] → [a] → [b]

Do a linear scan over the list, returning all successful classifications

data Event model cmd resp (r ∷ Type → Type)

Constructors

Event

Fields

eventBefore ∷ model r

eventCmd ∷ cmd r

eventAfter ∷ model r

eventResp ∷ resp r

execCmds ∷ ∀ model cmd m resp. StateMachine model cmd m resp → Commands cmd resp → [Event model cmd resp Symbolic]

execCmds is just the repeated form of execCmd.

execHistory ∷ ∀ model cmd m resp. StateMachine model cmd m resp → History cmd resp → [Event model cmd resp Concrete]