Test.StateMachine.Lockstep.NAry

Test type-level parameters

type family MockState t ∷ Type

data family Cmd t ∷ (Type → Type) → [Type] → Type

data family Resp t ∷ (Type → Type) → [Type] → Type

type family RealHandles t ∷ [Type]

data family MockHandle t a ∷ Type

type family RealMonad t ∷ Type → Type

type family Test (f ∷ (Type → Type) → [Type] → Type) ∷ Type where ...

Equations

Test (Cmd t) = t

Test (Resp t) = t

Test term-level parameters

data StateMachineTest t

Constructors

(Monad (RealMonad t), All Typeable (RealHandles t), All Eq (RealHandles t), All (And Show (Compose Show (MockHandle t))) (RealHandles t), All (And ToExpr (Compose ToExpr (MockHandle t))) (RealHandles t), NTraversable (Resp t), Eq (Resp t (MockHandle t) (RealHandles t)), Show (Resp t (MockHandle t) (RealHandles t)), Show (Resp t (FlipRef Symbolic) (RealHandles t)), Show (Resp t (FlipRef Concrete) (RealHandles t)), NTraversable (Cmd t), Show (Cmd t (FlipRef Symbolic) (RealHandles t)), Show (Cmd t (FlipRef Concrete) (RealHandles t)), Show (MockState t), ToExpr (MockState t)) ⇒ StateMachineTest

Fields

runMock ∷ Cmd t (MockHandle t) (RealHandles t) → MockState t → (Resp t (MockHandle t) (RealHandles t), MockState t)

runReal ∷ Cmd t I (RealHandles t) → RealMonad t (Resp t I (RealHandles t))

initMock ∷ MockState t

newHandles ∷ ∀ f. Resp t f (RealHandles t) → NP ([] :.: f) (RealHandles t)

generator ∷ Model t Symbolic → Maybe (Gen (Cmd t :@ Symbolic))

shrinker ∷ Model t Symbolic → (Cmd t :@ Symbolic) → [Cmd t :@ Symbolic]

cleanup ∷ Model t Concrete → RealMonad t ()

Handle instantiation

newtype At f r

Constructors

At

Fields

unAt ∷ f (FlipRef r) (RealHandles (Test f))

type (:@) f r = At f r

Model state

data Model t r

Constructors

Model

Fields

modelState ∷ MockState t

modelRefss ∷ Refss t r

newtype Refs t r a

Relation between real and mock references for single handle type a

Constructors

Refs

Fields

unRefs ∷ [(Reference a r, MockHandle t a)]

newtype Refss t r

Relation between real and mock references for all handle types

Constructors

Refss

Fields

unRefss ∷ NP (Refs t r) (RealHandles t)

newtype FlipRef r h

Constructors

FlipRef

Fields

unFlipRef ∷ Reference h r

Running the tests

prop_sequential

Arguments

∷ RealMonad t ~ IO

⇒ StateMachineTest t

→ Maybe Int
(Optional) minimum number of commands

→ Property

prop_parallel

Arguments

∷ RealMonad t ~ IO

⇒ StateMachineTest t

→ Maybe Int
(Optional) minimum number of commands

→ Property