Test.StateMachine.Markov

Description

This module contains helper functions for testing using Markov chains.

Documentation

data Markov state cmd_ prob

Markov chain.

makeMarkov ∷ Ord state ⇒ [Map state [Transition state cmd_ prob]] → Markov state cmd_ prob

Constructor for Markov chains.

toAdjacencyMap ∷ Ord state ⇒ Markov state cmd_ prob → Map state (Map state (cmd_, prob))

Expose inner graph structure of markov chain

(-<) ∷ Fractional prob ⇒ state → [Either (cmd_, state) ((cmd_, prob), state)] → Map state [Transition state cmd_ prob] infixl 5

Infix operator for starting to creating a transition in the Markov chain, finish the transition with one of '(>-)' or '(/-)' depending on whether the transition has a specific or a uniform probability.

(>-) ∷ (cmd_, prob) → state → Either (cmd_, state) ((cmd_, prob), state) infixl 5

Finish making a transition with a specified probability distribution.

(/-) ∷ cmd_ → state → Either (cmd_, state) ((cmd_, prob), state) infixl 5

Finish making a transition with an uniform probability distribution.

markovGenerator ∷ ∀ state cmd_ cmd model. (Show state, Show cmd_) ⇒ (Ord state, Ord cmd_) ⇒ Markov state cmd_ Double → Map cmd_ (model Symbolic → Gen (cmd Symbolic)) → (model Symbolic → state) → (state → Bool) → model Symbolic → Maybe (Gen (cmd Symbolic))

Create a generator from a Markov chain.

coverMarkov ∷ (Show state, Show cmd_, Testable prop) ⇒ Markov state cmd_ Double → prop → Property

Variant of QuickCheck's coverTable which works on Markov chains.

tabulateMarkov ∷ ∀ model state cmd cmd_ m resp prop. Testable prop ⇒ (Show state, Show cmd_) ⇒ StateMachine model cmd m resp → (model Symbolic → state) → (cmd Symbolic → cmd_) → Commands cmd resp → prop → Property

Variant of QuickCheck's tabulate which works for Markov chains.

transitionMatrix ∷ ∀ state cmd_. Ord state ⇒ (Generic state, GEnum FiniteEnum (Rep state), GBounded (Rep state)) ⇒ Markov state cmd_ Double → Matrix Double

stimulusMatrix ∷ ∀ state cmd. (Ord state, Ord cmd) ⇒ (Generic state, GEnum FiniteEnum (Rep state), GBounded (Rep state)) ⇒ (Generic cmd, GEnum FiniteEnum (Rep cmd), GBounded (Rep cmd)) ⇒ Markov state cmd Double → Matrix Double

historyObservations ∷ ∀ model cmd m resp state cmd_ prob. Ord state ⇒ Ord cmd_ ⇒ (Generic state, GEnum FiniteEnum (Rep state), GBounded (Rep state)) ⇒ StateMachine model cmd m resp → Markov state cmd_ prob → (model Concrete → state) → (cmd Concrete → cmd_) → History cmd resp → (Matrix Double, Matrix Double)

markovToDot ∷ (Show state, Show cmd_, Show prob) ⇒ state → state → Markov state cmd_ prob → String

markovToPs ∷ (Show state, Show cmd_, Show prob) ⇒ state → state → Markov state cmd_ prob → FilePath → IO ()

data StatsDb m

Constructors

StatsDb

Fields

store ∷ (Matrix Double, Matrix Double) → m ()

load ∷ m (Maybe (Matrix Double, Matrix Double))

type PropertyName = String

nullStatsDb ∷ Monad m ⇒ StatsDb m

fileStatsDb ∷ FilePath → PropertyName → StatsDb IO

persistStats ∷ Monad m ⇒ StatsDb m → (Matrix Double, Matrix Double) → PropertyM m ()

computeReliability ∷ Monad m ⇒ StatsDb m → Matrix Double → (Matrix Double, Matrix Double) → m (Double, Double)

printReliability ∷ Testable prop ⇒ StatsDb IO → Matrix Double → (Matrix Double, Matrix Double) → prop → Property

quickCheckReliability ∷ Testable prop ⇒ StatsDb IO → Matrix Double → prop → IO ()

testChainToDot ∷ ∀ state cmd_ prob m. (Show state, Ord state, Monad m) ⇒ (Generic state, GEnum FiniteEnum (Rep state)) ⇒ StatsDb m → state → state → Markov state cmd_ prob → m String