Test.StateMachine.Types.Environment

Description

This module contains environments that are used to translate between symbolic and concrete references. It's taken from the Hedgehog library.

Documentation

newtype Environment

A mapping of symbolic values to concrete values.

Constructors

Environment

Fields

unEnvironment ∷ Map Var Dynamic

data EnvironmentError

Environment errors.

Constructors

EnvironmentValueNotFound !Var

EnvironmentTypeError !TypeRep !TypeRep

emptyEnvironment ∷ Environment

Create an empty environment.

insertConcrete ∷ Var → Dynamic → Environment → Environment

Insert a symbolic / concrete pairing in to the environment.

insertConcretes ∷ [Var] → [Dynamic] → Environment → Environment

reifyDynamic ∷ ∀ a. Typeable a ⇒ Dynamic → Either EnvironmentError (Concrete a)

Cast a Dynamic in to a concrete value.

reifyEnvironment ∷ Environment → ∀ a. Symbolic a → Either EnvironmentError (Concrete a)

Turns an environment in to a function for looking up a concrete value from a symbolic one.

reify ∷ Traversable t ⇒ Environment → t Symbolic → Either EnvironmentError (t Concrete)

Convert a symbolic structure to a concrete one, using the provided environment.