Test.StateMachine.Types.History

Description

This module contains the notion of a history of an execution of a (parallel) program.

Documentation

newtype History cmd resp

Constructors

History

Fields

unHistory ∷ History' cmd resp

type History' cmd resp = [(Pid, HistoryEvent cmd resp)]

newtype Pid

Constructors

Pid

Fields

unPid ∷ Int

data HistoryEvent cmd resp

Constructors

Invocation !(cmd Concrete) !(Set Var)

Response !(resp Concrete)

Exception !String

data Operation cmd resp

An operation packs up an invocation event with its corresponding response event.

Constructors

Operation (cmd Concrete) (resp Concrete) Pid

Crash (cmd Concrete) String Pid

makeOperations ∷ History' cmd resp → [Operation cmd resp]

Given a sequential history, group invocation and response events into operations.

interleavings ∷ History' cmd resp → Forest (Operation cmd resp)

Given a parallel history, return all possible interleavings of invocations and corresponding response events.

operationsPath ∷ Forest (Operation cmd resp) → [Operation cmd resp]

completeHistory ∷ (cmd Concrete → resp Concrete) → History cmd resp → History cmd resp