Test.StateMachine.Z

Description

This module contains Z-inspried combinators for working with relations. The idea is that they can be used to define concise and showable models. This module is an experiment and will likely change or move to its own package.

Documentation

cons ∷ a → [a] → [a]

union ∷ Eq a ⇒ [a] → [a] → [a] infixr 6

intersect ∷ Eq a ⇒ [a] → [a] → [a] infixr 7

isSubsetOf ∷ (Eq a, Show a) ⇒ [a] → [a] → Logic infix 5

Subset.

>>> boolean ([1, 2] `isSubsetOf` [3, 2, 1])
True

(~=) ∷ (Eq a, Show a) ⇒ [a] → [a] → Logic infix 5

Set equality.

>>> boolean ([1, 1, 2] ~= [2, 1])
True

type Rel a b = [(a, b)]

Relations.

type Fun a b = Rel a b

(Partial) functions.

type (:<->) a b = Rel a b infixr 1

type (:->) a b = Fun a b infixr 1

type (:/->) a b = Fun a b infixr 1

empty ∷ Rel a b

identity ∷ [a] → Rel a a

singleton ∷ a → b → Rel a b

domain ∷ Rel a b → [a]

codomain ∷ Rel a b → [b]

compose ∷ Eq b ⇒ Rel b c → Rel a b → Rel a c

fcompose ∷ Eq b ⇒ Rel a b → Rel b c → Rel a c

inverse ∷ Rel a b → Rel b a

lookupDom ∷ Eq a ⇒ a → Rel a b → [b]

lookupCod ∷ Eq b ⇒ b → Rel a b → [a]

(<|) ∷ Eq a ⇒ [a] → Rel a b → Rel a b infixr 4

Domain restriction.

>>> ['a'] <| [ ('a', "apa"), ('b', "bepa") ]
[('a',"apa")]

(|>) ∷ Eq b ⇒ Rel a b → [b] → Rel a b infixl 4

Codomain restriction.

>>> [ ('a', "apa"), ('b', "bepa") ] |> ["apa"]
[('a',"apa")]

(<-|) ∷ Eq a ⇒ [a] → Rel a b → Rel a b infixr 4

Domain substraction.

>>> ['a'] <-| [ ('a', "apa"), ('b', "bepa") ]
[('b',"bepa")]

(|->) ∷ Eq b ⇒ Rel a b → [b] → Rel a b infixl 4

Codomain substraction.

>>> [ ('a', "apa"), ('b', "bepa"), ('c', "cepa") ] |-> ["apa"]
[('b',"bepa"),('c',"cepa")]

>>> [ ('a', "apa"), ('b', "bepa"), ('c', "cepa") ] |-> ["apa", "cepa"]
[('b',"bepa")]

>>> [ ('a', "apa"), ('b', "bepa"), ('c', "cepa") ] |-> ["apa"] |-> ["cepa"]
[('b',"bepa")]

image ∷ Eq a ⇒ Rel a b → [a] → [b]

The image of a relation.

(<+) ∷ (Eq a, Eq b) ⇒ Rel a b → Rel a b → Rel a b infixr 4

Overriding.

>>> [('a', "apa")] <+ [('a', "bepa")]
[('a',"bepa")]

>>> [('a', "apa")] <+ [('b', "bepa")]
[('a',"apa"),('b',"bepa")]

(<**>) ∷ Eq a ⇒ Rel a b → Rel a c → Rel a (b, c) infixl 4

Direct product.

(<||>) ∷ Rel a c → Rel b d → Rel (a, b) (c, d) infixl 4

Parallel product.

isTotalRel ∷ (Eq a, Show a) ⇒ Rel a b → [a] → Logic

isSurjRel ∷ (Eq b, Show b) ⇒ Rel a b → [b] → Logic

isTotalSurjRel ∷ (Eq a, Eq b, Show a, Show b) ⇒ Rel a b → [a] → [b] → Logic

isPartialFun ∷ (Eq a, Eq b, Show b) ⇒ Rel a b → Logic

isTotalFun ∷ (Eq a, Eq b, Show a, Show b) ⇒ Rel a b → [a] → Logic

isPartialInj ∷ (Eq a, Eq b, Show a, Show b) ⇒ Rel a b → Logic

isTotalInj ∷ (Eq a, Eq b, Show a, Show b) ⇒ Rel a b → [a] → Logic

isPartialSurj ∷ (Eq a, Eq b, Show b) ⇒ Rel a b → [b] → Logic

isTotalSurj ∷ (Eq a, Eq b, Show a, Show b) ⇒ Rel a b → [a] → [b] → Logic

isBijection ∷ (Eq a, Eq b, Show a, Show b) ⇒ Rel a b → [a] → [b] → Logic

(!) ∷ (Eq a, Show a, Show b) ⇒ Fun a b → a → b

Application.

(!?) ∷ Eq a ⇒ Fun a b → a → Maybe b

(.%) ∷ (Eq a, Eq b, Show a, Show b) ⇒ (Fun a b, a) → (b → b) → Fun a b infixr 4

(.!) ∷ Rel a b → a → (Rel a b, a) infixr 9

(.=) ∷ (Eq a, Eq b) ⇒ (Rel a b, a) → b → Rel a b infix 4

Assignment.

>>> singleton 'a' "apa" .! 'a' .= "bepa"
[('a',"bepa")]

>>> singleton 'a' "apa" .! 'b' .= "bepa"
[('a',"apa"),('b',"bepa")]