Test.StateMachine

Description

The main module for state machine based testing, it contains combinators that help you build sequential and parallel properties.

Sequential property combinators

forAllCommands

Arguments

∷ Testable prop

⇒ (Show (cmd Symbolic), Show (resp Symbolic), Show (model Symbolic))

⇒ (Traversable cmd, Foldable resp)

⇒ StateMachine model cmd m resp

→ Maybe Int
Minimum number of commands.

→ (Commands cmd resp → prop)
Predicate.

→ Property

existsCommands

Arguments

∷ (Testable prop, Foldable resp)

⇒ (Show (model Symbolic), Show (cmd Symbolic), Show (resp Symbolic))

⇒ StateMachine model cmd m resp

→ [model Symbolic → Gen (cmd Symbolic)]
Generators.

→ (Commands cmd resp → prop)
Predicate.

→ Property

Generate commands from a list of generators.

runCommands ∷ (Show (cmd Concrete), Show (resp Concrete)) ⇒ (Traversable cmd, Foldable resp) ⇒ (MonadMask m, MonadIO m) ⇒ StateMachine model cmd m resp → Commands cmd resp → PropertyM m (History cmd resp, model Concrete, Reason)

prettyCommands ∷ (MonadIO m, ToExpr (model Concrete)) ⇒ (Show (cmd Concrete), Show (resp Concrete)) ⇒ StateMachine model cmd m resp → History cmd resp → Property → PropertyM m ()

prettyCommands' ∷ (MonadIO m, ToExpr (model Concrete), ToExpr tag) ⇒ (Show (cmd Concrete), Show (resp Concrete)) ⇒ StateMachine model cmd m resp → ([Event model cmd resp Symbolic] → [tag]) → Commands cmd resp → History cmd resp → Property → PropertyM m ()

Variant of prettyCommands that also prints the tags covered by each command.

checkCommandNames ∷ ∀ cmd resp. CommandNames cmd ⇒ Commands cmd resp → Property → Property

Print the percentage of each command used. The prefix check is an unfortunate remaining for backwards compatibility.

coverCommandNames ∷ ∀ cmd resp. CommandNames cmd ⇒ Commands cmd resp → Property → Property

Fail if some commands have not been executed.

commandNames ∷ ∀ cmd resp. CommandNames cmd ⇒ Commands cmd resp → [(String, Int)]

commandNamesInOrder ∷ ∀ cmd resp. CommandNames cmd ⇒ Commands cmd resp → [String]

saveCommands ∷ (Show (cmd Symbolic), Show (resp Symbolic)) ⇒ FilePath → Commands cmd resp → Property → Property

runSavedCommands ∷ (Show (cmd Concrete), Show (resp Concrete)) ⇒ (Traversable cmd, Foldable resp) ⇒ (MonadMask m, MonadIO m) ⇒ (Read (cmd Symbolic), Read (resp Symbolic)) ⇒ StateMachine model cmd m resp → FilePath → PropertyM m (Commands cmd resp, History cmd resp, model Concrete, Reason)

showLabelledExamples ∷ (Show tag, Show (model Symbolic)) ⇒ (Show (cmd Symbolic), Show (resp Symbolic)) ⇒ (Traversable cmd, Foldable resp) ⇒ StateMachine model cmd m resp → ([Event model cmd resp Symbolic] → [tag]) → IO ()

showLabelledExamples'

Arguments

∷ (Show tag, Show (model Symbolic))

⇒ (Show (cmd Symbolic), Show (resp Symbolic))

⇒ (Traversable cmd, Foldable resp)

⇒ StateMachine model cmd m resp

→ Maybe Int
Seed

→ Int
Number of tests to run to find examples

→ ([Event model cmd resp Symbolic] → [tag])

→ (tag → Bool)
Tag filter (can be const True)

→ IO ()

Show minimal examples for each of the generated tags.

noCleanup ∷ Monad m ⇒ model Concrete → m ()

Parallel property combinators

forAllParallelCommands

Arguments

∷ Testable prop

⇒ (Show (cmd Symbolic), Show (resp Symbolic), Show (model Symbolic))

⇒ (Traversable cmd, Foldable resp)

⇒ StateMachine model cmd m resp

→ Maybe Int

→ (ParallelCommands cmd resp → prop)
Predicate.

→ Property

forAllNParallelCommands

Arguments

∷ Testable prop

⇒ (Show (cmd Symbolic), Show (resp Symbolic), Show (model Symbolic))

⇒ (Traversable cmd, Foldable resp)

⇒ StateMachine model cmd m resp

→ Int
Number of threads

→ (NParallelCommands cmd resp → prop)
Predicate.

→ Property

runNParallelCommands ∷ (Show (cmd Concrete), Show (resp Concrete)) ⇒ (Traversable cmd, Foldable resp) ⇒ (MonadMask m, MonadUnliftIO m) ⇒ StateMachine model cmd m resp → NParallelCommands cmd resp → PropertyM m [(History cmd resp, Logic)]

runParallelCommands ∷ (Show (cmd Concrete), Show (resp Concrete)) ⇒ (Traversable cmd, Foldable resp) ⇒ (MonadMask m, MonadUnliftIO m) ⇒ StateMachine model cmd m resp → ParallelCommands cmd resp → PropertyM m [(History cmd resp, Logic)]

runParallelCommands' ∷ (Show (cmd Concrete), Show (resp Concrete)) ⇒ (Traversable cmd, Foldable resp) ⇒ (MonadMask m, MonadUnliftIO m) ⇒ StateMachine model cmd m resp → (cmd Concrete → resp Concrete) → ParallelCommands cmd resp → PropertyM m [(History cmd resp, Logic)]

runParallelCommandsNTimes

Arguments

∷ (Show (cmd Concrete), Show (resp Concrete))

⇒ (Traversable cmd, Foldable resp)

⇒ (MonadMask m, MonadUnliftIO m)

⇒ Int
How many times to execute the parallel program.

→ StateMachine model cmd m resp

→ ParallelCommands cmd resp

→ PropertyM m [(History cmd resp, Logic)]

runNParallelCommandsNTimes'

Arguments

∷ (Show (cmd Concrete), Show (resp Concrete))

⇒ (Traversable cmd, Foldable resp)

⇒ (MonadMask m, MonadUnliftIO m)

⇒ Int
How many times to execute the parallel program.

→ StateMachine model cmd m resp

→ (cmd Concrete → resp Concrete)

→ NParallelCommands cmd resp

→ PropertyM m [(History cmd resp, Logic)]

runParallelCommandsNTimes'

Arguments

∷ (Show (cmd Concrete), Show (resp Concrete))

⇒ (Traversable cmd, Foldable resp)

⇒ (MonadMask m, MonadUnliftIO m)

⇒ Int
How many times to execute the parallel program.

→ StateMachine model cmd m resp

→ (cmd Concrete → resp Concrete)

→ ParallelCommands cmd resp

→ PropertyM m [(History cmd resp, Logic)]

runNParallelCommandsNTimes

Arguments

∷ (Show (cmd Concrete), Show (resp Concrete))

⇒ (Traversable cmd, Foldable resp)

⇒ (MonadMask m, MonadUnliftIO m)

⇒ Int
How many times to execute the parallel program.

→ StateMachine model cmd m resp

→ NParallelCommands cmd resp

→ PropertyM m [(History cmd resp, Logic)]

prettyNParallelCommands

Arguments

∷ (Show (cmd Concrete), Show (resp Concrete))

⇒ MonadIO m

⇒ Foldable cmd

⇒ NParallelCommands cmd resp

→ [(History cmd resp, Logic)]
Output of runNParallelCommands.

→ PropertyM m ()

prettyParallelCommands

Arguments

∷ (Show (cmd Concrete), Show (resp Concrete))

⇒ MonadIO m

⇒ Foldable cmd

⇒ ParallelCommands cmd resp

→ [(History cmd resp, Logic)]
Output of runNParallelCommands.

→ PropertyM m ()

prettyParallelCommandsWithOpts

Arguments

∷ (MonadIO m, Foldable cmd)

⇒ (Show (cmd Concrete), Show (resp Concrete))

⇒ ParallelCommands cmd resp

→ Maybe GraphOptions

→ [(History cmd resp, Logic)]
Output of runParallelCommands.

→ PropertyM m ()

Takes the output of parallel program runs and pretty prints a counterexample if any of the runs fail.

prettyNParallelCommandsWithOpts

Arguments

∷ (Show (cmd Concrete), Show (resp Concrete))

⇒ MonadIO m

⇒ Foldable cmd

⇒ NParallelCommands cmd resp

→ Maybe GraphOptions

→ [(History cmd resp, Logic)]
Output of runNParallelCommands.

→ PropertyM m ()

Takes the output of parallel program runs and pretty prints a counterexample if any of the runs fail.

checkCommandNamesParallel ∷ ∀ cmd resp t. Foldable t ⇒ CommandNames cmd ⇒ ParallelCommandsF t cmd resp → Property → Property

Print the percentage of each command used. The prefix check is an unfortunate remaining for backwards compatibility.

coverCommandNamesParallel ∷ ∀ cmd resp t. Foldable t ⇒ CommandNames cmd ⇒ ParallelCommandsF t cmd resp → Property → Property

Fail if some commands have not been executed.

commandNamesParallel ∷ ∀ cmd resp t. Foldable t ⇒ CommandNames cmd ⇒ ParallelCommandsF t cmd resp → [(String, Int)]

Types

data StateMachine model cmd m resp

Constructors

StateMachine (∀ r. model r) (∀ r. (Show1 r, Ord1 r) ⇒ model r → cmd r → resp r → model r) (model Symbolic → cmd Symbolic → Logic) (model Concrete → cmd Concrete → resp Concrete → Logic) (Maybe (model Concrete → Logic)) (model Symbolic → Maybe (Gen (cmd Symbolic))) (model Symbolic → cmd Symbolic → [cmd Symbolic]) (cmd Concrete → m (resp Concrete)) (model Symbolic → cmd Symbolic → GenSym (resp Symbolic)) (model Concrete → m ())

data Concrete a

data Symbolic a

data Reference a r

concrete ∷ Reference a Concrete → a

reference ∷ Typeable a ⇒ a → Reference a Concrete

newtype Opaque a

Constructors

Opaque

Fields

unOpaque ∷ a

opaque ∷ Reference (Opaque a) Concrete → a

data Reason

Constructors

Ok

PreconditionFailed String

PostconditionFailed String

InvariantBroken String

ExceptionThrown String

MockSemanticsMismatch

data GenSym a

genSym ∷ Typeable a ⇒ GenSym (Reference a Symbolic)

class CommandNames (cmd ∷ k → Type) where

The names of all possible commands

This is used for things like tagging, coverage checking, etc.

Minimal complete definition

Nothing

Methods

cmdName ∷ cmd r → String

Name of this particular command

cmdNames ∷ Proxy (cmd r) → [String]

Name of all possible commands

cmdName ∷ (Generic1 cmd, CommandNames (Rep1 cmd)) ⇒ cmd r → String

Name of this particular command

cmdNames ∷ ∀ r. CommandNames (Rep1 cmd) ⇒ Proxy (cmd r) → [String]

Name of all possible commands

module Test.StateMachine.Logic

module Test.StateMachine.Markov

Re-export

class ToExpr a

toExpr converts a Haskell value into untyped Haskell-like syntax tree, Expr.

>>> toExpr ((1, Just 2) :: (Int, Maybe Int))
App "_\215_" [App "1" [],App "Just" [App "2" []]]

toExpr ∷ ToExpr a ⇒ a → Expr